whee

+++ packages/hello-terrain/three/src/TerrainVaryings.ts
@@ -8,7 +8,10 @@ export class TerrainVaryings {
   readonly vGlobalVertexIndex: ReturnType<typeof varyingProperty>;
   readonly vElevation: ReturnType<typeof varyingProperty>;
   readonly vNormal: ReturnType<typeof varyingProperty>;
-  // Control map varyings for smooth interpolation across triangles
+  // Control map varyings:
+  // - IDs are DISCRETE, so keep them flat (int varyings) to avoid interpolating through
+  //   unrelated texture indices (causes visible "bands").
+  // - Blend is continuous, so keep it interpolated (f32).
   readonly vControlBaseId: ReturnType<typeof varyingProperty>;
   readonly vControlOverlayId: ReturnType<typeof varyingProperty>;
   readonly vControlBlend: ReturnType<typeof varyingProperty>;
@@ -24,12 +27,13 @@ export class TerrainVaryings {
     );
     this.vElevation = varyingProperty("f32", `vElevation${suffix}`);
     this.vNormal = varyingProperty("vec3", `vNormal${suffix}`);
-    // Control map values as floats for smooth interpolation
-    this.vControlBaseId = varyingProperty("f32", `vControlBaseId${suffix}`);
+    // Texture IDs: int (flat)
+    this.vControlBaseId = varyingProperty("int", `vControlBaseId${suffix}`);
     this.vControlOverlayId = varyingProperty(
-      "f32",
+      "int",
       `vControlOverlayId${suffix}`
     );
+    // Blend factor: f32 (interpolated)
     this.vControlBlend = varyingProperty("f32", `vControlBlend${suffix}`);
   }
 }
